[REFACTOR] Move URL arguments loading to a special file
